refactor slugOrId
