+++ server/api/experiments/[slug].get.ts
@@ -1,15 +1,6 @@
-import { validate as uuidValidate } from "uuid"
-
 export default defineEventHandler(async (event) => {
-  const slugOrId = getRouterParam(event, "slug")
-  if (!slugOrId) {
-    throw createError({ status: 400, message: "Invalid slug" })
-  }
-
-  const isId = uuidValidate(slugOrId)
-  const whereClause = isId ? { id: slugOrId } : { slug: slugOrId }
   const experiment = await prisma.experiment.findFirst({
-    where: whereClause,
+    where: getSlugOrIdPrismaWhereClause(event),
   })
 
   if (!experiment) {
