+++ server/utils/utils.ts
@@ -1,3 +1,6 @@
+import { validate as uuidValidate } from "uuid"
+import type { H3Event, EventHandlerRequest } from "h3"
+
 /**
  * Consumes a value of type T without performing any operations.
  * This is a no-op function that essentially ignores its input.
@@ -6,4 +9,18 @@
  */
 export function consume<T>(_: T) {
   // empty
+}
+
+/**
+ * Generates a Prisma where clause based on a slug or ID parameter from the event.
+ */
+export function getSlugOrIdPrismaWhereClause(event: H3Event<EventHandlerRequest>) {
+  const slugOrId = getRouterParam(event, "slug")
+  if (!slugOrId) {
+    throw createError({ status: 400, message: "Invalid slug" })
+  }
+
+  const isId = uuidValidate(slugOrId)
+  const whereClause = isId ? { id: slugOrId } : { slug: slugOrId }
+  return whereClause
 }
